+++ docs/references/index.rst
@@ -9,7 +9,7 @@ The package is split up into submodules.
     :maxdepth: 2
 
     data
-    brain
+    localization
     features
     array_ops
     encoding
