+++ naplib/__init__.py
@@ -38,11 +38,11 @@ import naplib.visualization
 import naplib.encoding
 import naplib.io
 import naplib.preprocessing
+import naplib.localization
 import naplib.array_ops
 import naplib.model_selection
 import naplib.utils
 from .data import Data, join_fields, concat
-from .brain import Brain
 import naplib.naplab
 
 __version__ = "2.0.0"
