+++ naplib/localization/__init__.py
@@ -1,0 +1,3 @@
+from .freesurfer import Brain, find_closest_vertices
+
+__all__ = ['Brain', 'find_closest_vertices']
