+++ naplib/visualization/__init__.py
@@ -1,5 +1,6 @@
 from .plots import shaded_error_plot, hierarchical_cluster_plot, strf_plot, freq_response, kde_plot
 from .eeg_cap_positions import eeg_locs
+from .brain_plots import plot_brain_elecs, plot_brain_overlay
 
-__all__  = ['eeg_locs', 'shaded_error_plot', 'hierarchical_cluster_plot', 'strf_plot', 'freq_response', 'kde_plot']
+__all__  = ['eeg_locs', 'shaded_error_plot', 'hierarchical_cluster_plot', 'strf_plot', 'freq_response', 'kde_plot', 'plot_brain_elecs', 'plot_brain_overlay']
